+++ lang/en/settings.php
@@ -5,6 +5,7 @@
  * @license    GPL 2 (http://www.gnu.org/licenses/gpl.html)
  * @author     Terence J. Grant<tjgrant@tatewake.com>
  */
- 
+
 $lang['cite_publisher'] = 'Publisher name (default is Wiki Title)';
 $lang['cite_author'] = 'Author name for pages (default is "Anonymous Contributors")';
+$lang['showcitebutton'] = 'Show Cite button (only when supported by your template)';