+++ lang/nl/settings.php
@@ -1,0 +1,11 @@
+<?php
+/**
+ * Dutch language file
+ *
+ * @license    GPL 2 (http://www.gnu.org/licenses/gpl.html)
+ * @author     G. Uitslag<klapinklapin@gmail.com>
+ */
+
+$lang['cite_publisher'] = 'Naam van de uitgever (standaard is dat de Wiki titel)';
+$lang['cite_author'] = 'Auteurnaam voor de pagina\'s (standaard is "Anonymous Contributors")';
+$lang['showcitebutton'] = 'Laat de Citaties knop zien (alleen als je template dit ondersteunt)';